+++ iti-react/Changelog.md
@@ -448,4 +448,5 @@ const myTab: Tab = ['tabName', 'Tab Label', { className: 'my-tab' }]
     -   Deprecate the old onChildReady
     -   Add onChildReady2. See OnChildReady.ts for the migration path
 -   \*\*\* rename ActionDialogProps.loading to actionInProgress
+-   Bring in FormCheck
 -   Add `validated-input` class to the `<div>` that wraps ValidatedInput, .etc